chore: tidy: move helm util code
Move helm utility functions common/k8stest
